Fixes authenticaton status due to API change
At some point the Nessus API changed its order and the 'status' node may
not be the first node in a reply. The utils.get_text_by_tag would return
the result of the first 'status' childNode, even if it was in a contents
nodeName.

This change runs through the reply node searching for a status nodeName:

In [46]: reply.childNodes
Out[46]:
[<DOM Element: contents at 0x10c8b2998>,
 <DOM Text node "u'\n'">,
 <DOM Element: status at 0x10c911050>,
 <DOM Text node "u'\n'">]